refactor : 클라리언트 전용 hydration 가드 추가
